requirements_no_tqdm.txt
Flask
dotenv
google-ai-generativelanguage==0.6.15
google-api-core
google-api-python-client
google-auth>=2.15.0
protobuf
requests<3.0.0,>=2.18.0
grpcio<2.0.0,>=1.33.2
grpcio-status<2.0.0,>=1.33.2
cachetools<6.0,>=2.0.0
pyasn1-modules>=0.2.1
rsa<5,>=3.1.4
charset-normalizer<4,>=2
idna<4,>=2.5
urllib3<3,>=1.21.1
certifi>=2017.4.17
pyasn1>=0.1.3
httplib2<1.0.0,>=0.19.0
google-auth-httplib2<1.0.0,>=0.2.0
uritemplate<5,>=3.0.1
pyparsing!=3.0.0,!=3.0.1,!=3.0.2,!=3.0.3,<4,>=2.4.2
proto-plus<2.0.0dev,>=1.22.3